Fix overflow in data.

+++ include/tiled-reader/layer.h
@@ -15,7 +15,7 @@ namespace tiled {
 class Map;
 
 struct TileIndex {
-	TileIndex(uint32_t data);
+	TileIndex(uint64_t data);
 
 	uint32_t gid:29;
 	bool flipped_horizontally:1;
